Refactor functions to their own files
