+++ src/testData/colors.ts
@@ -1,0 +1,122 @@
+// Known HEX Values
+export const HEX_TEAL = '#3cd6bf';
+export const HEX_WHITE = '#ffffff';
+export const HEX_BLACK = '#000000';
+export const HEX_RED = '#ff0000';
+export const HEX_YELLOW = '#ffff00';
+export const HEX_GREEN = '#00ff00';
+export const HEX_CYAN = '#00ffff';
+export const HEX_BLUE = '#0000ff';
+export const HEX_MAGENTA = '#ff00ff';
+
+
+// Known RGB Values
+export const RGB_TEAL = {
+  red: 60,
+  green: 214,
+  blue: 191,
+};
+
+export const RGB_WHITE = {
+  red: 255,
+  green: 255,
+  blue: 255,
+};
+
+export const RGB_BLACK = {
+  red: 0,
+  green: 0,
+  blue: 0,
+};
+
+export const RGB_RED = {
+  red: 255,
+  green: 0,
+  blue: 0,
+};
+
+export const RGB_YELLOW = {
+  red: 255,
+  green: 255,
+  blue: 0,
+};
+
+export const RGB_GREEN = {
+  red: 0,
+  green: 255,
+  blue: 0,
+};
+
+export const RGB_CYAN = {
+  red: 0,
+  green: 255,
+  blue: 255,
+};
+
+export const RGB_BLUE = {
+  red: 0,
+  green: 0,
+  blue: 255,
+};
+
+export const RGB_MAGENTA = {
+  red: 255,
+  green: 0,
+  blue: 255,
+};
+
+
+// Known HSV Values
+export const HSV_TEAL = {
+  hue: 171.03896103896102,
+  saturation: 0.7196261682242991,
+  value: 0.8392156862745098,
+};
+
+export const HSV_WHITE = {
+  hue: 0,
+  saturation: 0,
+  value: 1,
+};
+
+export const HSV_BLACK = {
+  hue: 0,
+  saturation: 0,
+  value: 0,
+};
+
+export const HSV_RED = {
+  hue: 0,
+  saturation: 1,
+  value: 1,
+};
+
+export const HSV_YELLOW = {
+  hue: 60,
+  saturation: 1,
+  value: 1,
+};
+
+export const HSV_GREEN = {
+  hue: 120,
+  saturation: 1,
+  value: 1,
+};
+
+export const HSV_CYAN = {
+  hue: 180,
+  saturation: 1,
+  value: 1,
+};
+
+export const HSV_BLUE = {
+  hue: 240,
+  saturation: 1,
+  value: 1,
+};
+
+export const HSV_MAGENTA = {
+  hue: 300,
+  saturation: 1,
+  value: 1,
+};
