+++ tsconfig.json
@@ -12,5 +12,9 @@
   },
   "include": [
     "src"
+  ],
+  "exclude": [
+    "src/**/*.test.ts",
+    "src/testData"
   ]
 }
